Remove unused pool/CPU mining option.

+++ src/rpcblockchain.cpp
@@ -3500,8 +3500,7 @@ Array MagnitudeReport(std::string cpid)
                                                 StructCPID stCPID = GetLifetimeCPID(structMag.cpid,"MagnitudeReport");
                                                 double days = (GetAdjustedTime() - stCPID.LowLockTime) / 86400.0;
                                                 entry.push_back(Pair("CPID",structMag.cpid));
-                                                // entry.push_back(Pair("PoolMining",bPoolMiningMode));
-                                                double dWeight = (double)GetRSAWeightByCPID(structMag.cpid);
+                                                double dWeight = GetRSAWeightByCPID(structMag.cpid);
                                                 //entry.push_back(Pair("RSA Weight",dWeight));
                                                 StructCPID UH = GetInitializedStructCPID2(cpid,mvMagnitudes);
                                                 // entry.push_back(Pair("RSA block count",UH.Accuracy));
